Move localize parsing into localize root folder
