+++ src/ObjLoading/Localize/Parsing/LocalizeFileParser.cpp
