+++ src/ObjLoading/Localize/Parsing/LocalizeFileParser.h
