+++ src/ObjLoading/Localize/Parsing/LocalizeFileParserState.cpp
