+++ src/ObjLoading/Localize/Parsing/LocalizeFileParserState.h
