+++ src/ObjLoading/Localize/Parsing/LocalizeFileReader.cpp
