+++ src/ObjLoading/Localize/Parsing/LocalizeFileReader.h
