+++ src/ObjLoading/Localize/Parsing/Sequence/SequenceLocalizeFileConsumeEmptyLines.cpp
