+++ src/ObjLoading/Localize/Parsing/Sequence/SequenceLocalizeFileEndMarker.cpp
