+++ src/ObjLoading/Localize/Parsing/Sequence/SequenceLocalizeFileNotes.cpp
